HOC to wrap with column preferences provider
