+++ apps/modernization-ui/src/design-system/table/preferences/index.ts
@@ -1,3 +1,5 @@
-export { useColumnPreferences, ColumnPreferenceProvider } from './useColumnPreferences';
-export type { ColumnPreference } from './useColumnPreferences';
+export { useColumnPreferences, useMaybeColumnPreferences } from './useColumnPreferences';
+export type { ColumnPreference } from './ColumnPreferenceProvider';
 export { ColumnPreferencesPanel } from './ColumnPreferencesPanel';
+export { ColumnPreferenceProvider } from './ColumnPreferenceProvider';
+export { withColumnPreferences } from './withColumnPreferences';
